renaming NsOptions::Boolean option handler class
* less typing FTW
